# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MedConsult API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List customers
      parameters:
        - name: agencyId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Customers list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Register/create customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerBody"
      responses:
        "201":
          description: Created customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Not found
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCustomerBody"
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # --- QUESTIONNAIRES ---
...
  /consultation-records:
    get:
      operationId: listConsultationRecords
      tags: [consultation-records]
      summary: List consultation records
      parameters:
        - name: appointmentId
          in: query
          schema:
            type: ["integer", "null"]
        - name: customerId
          in: query
          schema:
            type: ["integer", "null"]
        - name: surgeonId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Consultation records list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConsultationRecord"
    post:
      operationId: createConsultationRecord
      tags: [consultation-records]
      summary: Create consultation record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConsultationRecordBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsultationRecord"
  /consultation-records/{id}:
    get:
      operationId: getConsultationRecord
      tags: [consultation-records]
      summary: Get consultation record by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Consultation record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsultationRecord"
        "404":
          description: Not found
    patch:
      operationId: updateConsultationRecord
      tags: [consultation-records]
      summary: Update consultation record (notes, answers, status)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateConsultationRecordBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsultationRecord"
  /consultation-records/{id}/complete:
    post:
      operationId: completeConsultationRecord
      tags: [consultation-records]
      summary: Mark consultation as completed and signed off
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsultationRecord"

  # --- MEDIA ---
  /consultation-records/{recordId}/media:
    get:
      operationId: listConsultationMedia
      tags: [media]
      summary: List media files for a consultation record
      parameters:
        - name: recordId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Media list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConsultationMedia"
    post:
      operationId: addConsultationMedia
      tags: [media]
      summary: Add media file reference to a consultation record
      parameters:
        - name: recordId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConsultationMediaBody"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsultationMedia"
  /consultation-records/{recordId}/media/{id}:
    delete:
      operationId: deleteConsultationMedia
      tags: [media]
      summary: Delete media file from consultation record
      parameters:
        - name: recordId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # --- STORAGE ---
...
components:
  schemas:
...
    Customer:
      type: object
      properties:
        id:
          type: integer
        agencyId:
          type: integer
        userId:
          type: ["integer", "null"]
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        dialingCode:
          type: ["string", "null"]
        nationality:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        postcode:
          type: ["string", "null"]
        preferredLanguage:
          type: ["string", "null"]
        medicalServicesInterest:
          type: ["string", "null"]
        heightCm:
          type: ["number", "null"]
        weightKg:
          type: ["number", "null"]
        heightUnit:
          type: ["string", "null"]
          enum: [cm, ft, "null"]
        weightUnit:
          type: ["string", "null"]
          enum: [kg, lbs, "null"]
        declarationSigned:
          type: boolean
        declarationSignedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - agencyId
        - firstName
        - lastName
        - declarationSigned
        - createdAt
        - updatedAt
    CreateCustomerBody:
      type: object
      properties:
        agencyId:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        dialingCode:
          type: ["string", "null"]
        nationality:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        postcode:
          type: ["string", "null"]
        preferredLanguage:
          type: ["string", "null"]
        medicalServicesInterest:
          type: ["string", "null"]
        heightCm:
          type: ["number", "null"]
        weightKg:
          type: ["number", "null"]
        heightUnit:
          type: ["string", "null"]
        weightUnit:
          type: ["string", "null"]
      required:
        - agencyId
        - firstName
        - lastName
    UpdateCustomerBody:
      type: object
      properties:
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        dialingCode:
          type: ["string", "null"]
        nationality:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        postcode:
          type: ["string", "null"]
        preferredLanguage:
          type: ["string", "null"]
        medicalServicesInterest:
          type: ["string", "null"]
        heightCm:
          type: ["number", "null"]
        weightKg:
          type: ["number", "null"]
        heightUnit:
          type: ["string", "null"]
        weightUnit:
          type: ["string", "null"]
        declarationSigned:
          type: ["boolean", "null"]
        declarationSignedAt:
          type: ["string", "null"]
...
    ConsultationRecord:
      type: object
      properties:
        id:
          type: integer
        appointmentId:
          type: integer
        surgeonId:
          type: integer
        customerId:
          type: integer
        notes:
          type: ["string", "null"]
        surgeonAnswers:
          type: ["string", "null"]
        status:
          type: string
          enum: [in_progress, completed]
        completedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - appointmentId
        - surgeonId
        - customerId
        - status
        - createdAt
        - updatedAt
    CreateConsultationRecordBody:
      type: object
      properties:
        appointmentId:
          type: integer
        surgeonId:
          type: integer
        customerId:
          type: integer
        notes:
          type: ["string", "null"]
        surgeonAnswers:
          type: ["string", "null"]
      required:
        - appointmentId
        - surgeonId
        - customerId
    UpdateConsultationRecordBody:
      type: object
      properties:
        notes:
          type: ["string", "null"]
        surgeonAnswers:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
          enum: [in_progress, completed, "null"]
    ConsultationMedia:
      type: object
      properties:
        id:
          type: integer
        consultationRecordId:
          type: integer
        mediaType:
          type: string
          enum: [photo, document, voice_recording]
        fileName:
          type: string
        objectKey:
          type: string
        url:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - consultationRecordId
        - mediaType
        - fileName
        - objectKey
        - createdAt
    CreateConsultationMediaBody:
      type: object
      properties:
        mediaType:
          type: string
          enum: [photo, document, voice_recording]
        fileName:
          type: string
        objectKey:
          type: string
      required:
        - mediaType
        - fileName
        - objectKey
...